+++ api/entrypoint.sh
@@ -1,8 +1,4 @@
 #!/bin/bash
-cd /sourcemaps_reader
-nohup node server.js &> /tmp/sourcemaps_reader.log &
-cd /assist_server
-nohup node server.js &> /tmp/assist_server.log &
-cd ..
+cd /
 python env_handler.py
 chalice local --no-autoreload --host 0.0.0.0 --stage ${ENTERPRISE_BUILD}
