--- api/sourcemaps_reader/.gitignore
@@ -1,11 +1,0 @@
-# package directories
-node_modules
-jspm_packages
-
-# Serverless directories
-.serverless/*.zip
-
-
-node_modules/
-.idea
-test.js
