--- api/sourcemaps_reader/README.md
@@ -1,15 +1,0 @@
-# sourcemap-reader
-Source Map Reader
-
-# For SAAS:
-to run local; put your test values in handler then run `node handler.js`
-to deploy `sls deploy --stage [staging|prod|dev]`
-
-# Requirements:
-- nodeJS 12 or greater
-
-# Install for OS:
-```
-npm install
-node server.js
-```
