--- api/sourcemaps_reader/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "sourcemap-reader",
-  "version": "1.0.0",
-  "description": "",
-  "main": "handler.js",
-  "dependencies": {
-    "aws-sdk": "^2.654.0",
-    "source-map": "^0.7.3"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Kraiem taha yassine <tahayk2@gmail.com>",
-  "license": "ISC"
-}
